refactor: rename contents directories to components
- Rename privacy/contents/ → privacy/components/
- Rename agreement/contents/ → agreement/components/
- Update import statements in page.tsx files

This follows the standard convention where component-related content
is organized under components/ directories.
