+++ src/cook-web/src/app/agreement/components/en-agreement.mdx
