+++ src/cook-web/src/app/agreement/components/zh-cn-agreement.mdx
