+++ src/cook-web/src/app/privacy/components/en-privacy.mdx
