+++ src/cook-web/src/app/privacy/components/zh-cn-privacy.mdx
